Build: Cleaning up the build and adding an initial version of the install script  - Moved all non-public headers to Private sub-folder
Former-commit-id: eb4b6fa6e468726c68d581effd400e1444ec9308

+++ Source/BansheeEngine/Renderer/BsRendererMaterial.h
@@ -125,7 +125,7 @@ namespace bs { namespace ct
 		/** Retrieves an instance of a particular variation of this renderer material. */
 		static T* get(const ShaderVariation& variation)
 		{
-			if(variation.getIdx() == -1)
+			if(variation.getIdx() == (UINT32)-1)
 				variation.setIdx(mMetaData.variations.find(variation));
 
 			UINT32 varIdx = variation.getIdx();
